Change score type in Leaderboard from int to long

+++ Assets/Game/Code/GamePlatformSdk/Runtime/YandexGames/YandexGamesCloudSave.cs
@@ -40,7 +40,7 @@ namespace YellowSquad.GamePlatformSdk
             PlayerPrefs.SetString(leaderboardName, value.ToString());
             
             if (PlayerAccount.IsAuthorized)
-                Leaderboard.SetScore(leaderboardName, (int)value);
+                Leaderboard.SetScore(leaderboardName, value);
         }
 
         public void Save()
